Vizia/plugins/vizia-ocr/requirements.txt
pytesseract==0.3.10
deep-translator==1.11.4
requests==2.31.0
PyMuPDF==1.23.8
Pillow==10.2.0